refactor(stats): move the legacy stats SQL oracle into a _test.go file
After the badge port, every production Database stats method delegates to the
storage backend; the legacy* hand-written SQL implementations and their
helpers (pr, snowieER, build*WhereClause, scanPositionIDs, the statsBaseJoin/
statsErrExpr/statsCountedExpr fragments) are referenced only by the parity
tests, where they serve as the differential oracle the storage backend is
pinned against.

Relocate them verbatim from db_stats.go into stats_legacy_oracle_test.go.
Pure code move, no behaviour change: db_stats.go shrinks 1179 → 293 lines and
now contains only the public types and the storage-delegating methods, so the
duplicate SQL no longer ships in the binary while the parity net is fully
preserved.

Verified: go build ./..., go vet, golangci-lint (0 issues), and
`go test -race ./pkg/blunderdb/database/` all green (the parity tests still
compile against and exercise the relocated oracle).
