+++ pkg/blunderdb/database/db_stats.go
@@ -2,9 +2,6 @@ package database
 
 import (
 	"context"
-	"fmt"
-	"math"
-	"strings"
 )
 
 // StatsDateRange holds the earliest and latest match dates in the database.
@@ -23,50 +20,6 @@ func (d *Database) GetStatsDateRange() StatsDateRange {
 	r, _ := d.store.Stats().DateRange(context.Background(), "")
 	return fromStorageDateRange(r)
 }
-
-func legacyGetStatsDateRange(d *Database) StatsDateRange {
-	d.mu.RLock()
-	defer d.mu.RUnlock()
-	if d.db == nil {
-		return StatsDateRange{}
-	}
-	var min, max string
-	_ = d.db.QueryRow(
-		`SELECT COALESCE(MIN(SUBSTR(match_date,1,10)),''), COALESCE(MAX(SUBSTR(match_date,1,10)),'')
-		 FROM match
-		 WHERE match_date IS NOT NULL AND match_date != '' AND match_date != '0001-01-01T00:00:00Z'`,
-	).Scan(&min, &max)
-	return StatsDateRange{DateFrom: min, DateTo: max}
-}
-
-// statsErrExpr is the SQL CASE expression that selects the correct error column
-// based on position decision type. Shared with db_search.go.
-const statsErrExpr = "CASE WHEN p.decision_type = 1 THEN a.cube_error ELSE a.best_move_equity_error END"
-
-// blunderThresholdMP is the error threshold (in stored millipoints units) above
-// which a decision is counted as a blunder. 100 ≈ 0.1 EMG.
-const blunderThresholdMP = 100
-
-// statsCountedExpr is the SQL predicate selecting the decisions that count
-// toward PR and decision tallies (XG semantics):
-//   - Checker: unforced positions only (a.is_forced = 0).
-//   - Cube: a position is counted when either (a) it is a "close" decision per
-//     gnuBG isCloseCubedecision (a.is_close_cube = 1) with the exclusion below,
-//     OR (b) the player took an active cube action other than NoDouble (Double,
-//     Take, Pass). The OR ensures premature doublings — wrong doubles with a
-//     large equity gap that are technically "not close" — are still counted.
-//
-// Exclusion for close NoDoubles: XG's stored EMG equities at extreme match
-// scores (mover's away ≤ 2 with centered cube) are amplified by a factor of
-// ~3–4× beyond the normal [-1, 1] range, so gnuBG's 0.16 threshold falsely
-// marks these positions as "close". We exclude correctly-played (cube_error=0)
-// NoDouble positions where the cube is still centered (cube_value=0) and the
-// mover needs ≤ 2 points to win, mirroring XG's actual decision counting.
-// Active cube actions (Double, Take, Pass) and missed doubles (cube_error > 0)
-// at this score are always counted regardless.
-//
-// Note: mv is accessible because statsBaseJoin includes JOIN move mv.
-const statsCountedExpr = "((p.decision_type = 0 AND a.is_forced = 0) OR (p.decision_type = 1 AND (COALESCE(mv.cube_action, '') NOT IN ('', 'No Double', 'NoDouble') OR (a.is_close_cube = 1 AND NOT (COALESCE(a.cube_error, 0) = 0 AND COALESCE(p.cube_value, 0) = 0 AND CASE WHEN mv.player = 1 THEN COALESCE(p.score_1, 99) ELSE COALESCE(p.score_2, 99) END <= 2)))))"
 
 // StatsFilter defines the filtering criteria for ComputeStats.
 type StatsFilter struct {
@@ -155,96 +108,6 @@ type StatsResult struct {
 	TopBlunders         []BlunderEntry    `json:"TopBlunders"`
 }
 
-// pr computes the Performance Rating from a sum of errors (millipoints stored
-// units) and the number of decisions. Formula: 500 × sumErrMP / 1000 / nDecisions.
-func pr(sumErrMP int64, nDecisions int) float64 {
-	if nDecisions == 0 {
-		return 0
-	}
-	return 500 * float64(sumErrMP) / 1000 / float64(nDecisions)
-}
-
-// snowieER computes the Snowie Error Rate from a sum of errors (millipoints
-// stored units) and the total checker move count for both players combined.
-// Formula identical to pr() but denominator = anTotalMoves[P1] + anTotalMoves[P2]
-// (gnubg/formatgs.c:415-424). The per-player metric is asymmetric: each player's
-// error is divided by the combined move count of both players, matching gnuBG.
-func snowieER(sumErrMP int64, nMovesBoth int) float64 {
-	if nMovesBoth == 0 {
-		return 0
-	}
-	return 500 * float64(sumErrMP) / 1000 / float64(nMovesBoth)
-}
-
-// statsBaseJoin is the FROM + JOIN fragment shared by all stats queries.
-const statsBaseJoin = `FROM position p
-JOIN analysis a ON a.position_id = p.id
-JOIN move mv ON mv.position_id = p.id
-JOIN game g ON g.id = mv.game_id
-JOIN match m ON m.id = g.match_id
-LEFT JOIN tournament t ON t.id = m.tournament_id`
-
-// buildBaseWhereClause constructs the base WHERE clause for the given filter
-// without the statsCountedExpr predicate. Used by buildStatsWhereClause (PR
-// queries) and by Snowie ER queries that need all moves (forced checker and
-// non-close cube included).
-func buildBaseWhereClause(filter StatsFilter) (whereSQL string, args []any) {
-	var clauses []string
-
-	if filter.PlayerName != "" {
-		clauses = append(clauses, "((m.player1_name = ? AND mv.player = 1) OR (m.player2_name = ? AND mv.player = -1))")
-		args = append(args, filter.PlayerName, filter.PlayerName)
-	}
-
-	if len(filter.TournamentIDs) > 0 {
-		placeholders := strings.Repeat("?,", len(filter.TournamentIDs))
-		placeholders = placeholders[:len(placeholders)-1]
-		clauses = append(clauses, "m.tournament_id IN ("+placeholders+")")
-		for _, id := range filter.TournamentIDs {
-			args = append(args, id)
-		}
-	}
-
-	if filter.DateFrom != "" && filter.DateTo != "" {
-		clauses = append(clauses, "m.match_date BETWEEN ? AND ?")
-		args = append(args, filter.DateFrom, filter.DateTo)
-	} else if filter.DateFrom != "" {
-		clauses = append(clauses, "m.match_date >= ?")
-		args = append(args, filter.DateFrom)
-	} else if filter.DateTo != "" {
-		clauses = append(clauses, "m.match_date <= ?")
-		args = append(args, filter.DateTo)
-	}
-
-	if filter.DecisionType >= 0 {
-		clauses = append(clauses, "p.decision_type = ?")
-		args = append(args, filter.DecisionType)
-	}
-
-	if len(filter.MatchLength) > 0 {
-		placeholders := strings.Repeat("?,", len(filter.MatchLength))
-		placeholders = placeholders[:len(placeholders)-1]
-		clauses = append(clauses, "m.match_length IN ("+placeholders+")")
-		for _, ml := range filter.MatchLength {
-			args = append(args, ml)
-		}
-	}
-
-	clauses = append(clauses, "a.position_id IS NOT NULL")
-	clauses = append(clauses, "("+statsErrExpr+") IS NOT NULL")
-
-	whereSQL = " WHERE " + strings.Join(clauses, " AND ")
-	return whereSQL, args
-}
-
-// buildStatsWhereClause wraps buildBaseWhereClause and appends the
-// statsCountedExpr predicate (XG/gnuBG semantics: unforced checker + close cube).
-func buildStatsWhereClause(filter StatsFilter) (whereSQL string, args []any) {
-	whereSQL, args = buildBaseWhereClause(filter)
-	whereSQL += " AND " + statsCountedExpr
-	return whereSQL, args
-}
-
 // ComputeStats aggregates performance metrics for the given filter.
 // ComputeStats delegates to the storage StatsStore (the single production
 // implementation, shared with the headless server). legacyComputeStats keeps
@@ -258,375 +121,6 @@ func (d *Database) ComputeStats(filter StatsFilter) (*StatsResult, error) {
 	}
 	return fromStorageStatsResult(r), nil
 }
-
-func legacyComputeStats(d *Database, filter StatsFilter) (*StatsResult, error) {
-	d.mu.RLock()
-	defer d.mu.RUnlock()
-
-	if d.db == nil {
-		return nil, fmt.Errorf("no database is currently open")
-	}
-
-	whereSQL, baseArgs := buildStatsWhereClause(filter)
-
-	result := &StatsResult{
-		PRRolling: make(map[int]float64),
-	}
-
-	// ── 1. Totals ────────────────────────────────────────────────────────────
-	row := d.db.QueryRow(
-		`SELECT COUNT(DISTINCT p.id), COUNT(DISTINCT m.id), COUNT(DISTINCT m.tournament_id), COUNT(*) `+
-			statsBaseJoin+whereSQL,
-		baseArgs...,
-	)
-	if err := row.Scan(
-		&result.Totals.NumPositions,
-		&result.Totals.NumMatches,
-		&result.Totals.NumTournaments,
-		&result.Totals.NumDecisions,
-	); err != nil {
-		return nil, fmt.Errorf("totals query: %w", err)
-	}
-
-	// ── 2. PR global + per decision_type ─────────────────────────────────────
-	rows, err := d.db.Query(
-		`SELECT p.decision_type, SUM(`+statsErrExpr+`), COUNT(*) `+
-			statsBaseJoin+whereSQL+
-			` GROUP BY p.decision_type`,
-		baseArgs...,
-	)
-	if err != nil {
-		return nil, fmt.Errorf("PR by decision_type query: %w", err)
-	}
-	var totalErrSum int64
-	var totalErrCount int
-	func() {
-		defer rows.Close()
-		for rows.Next() {
-			var dt int
-			var sumErr int64
-			var cnt int
-			if err2 := rows.Scan(&dt, &sumErr, &cnt); err2 != nil {
-				return
-			}
-			totalErrSum += sumErr
-			totalErrCount += cnt
-			switch dt {
-			case 0:
-				result.PRChecker = pr(sumErr, cnt)
-			case 1:
-				result.PRCube = pr(sumErr, cnt)
-			}
-		}
-	}()
-	result.PRGlobal = pr(totalErrSum, totalErrCount)
-
-	// ── Snowie ER (global) ────────────────────────────────────────────────────
-	// Uses buildBaseWhereClause (no statsCountedExpr): all moves including forced
-	// checker and non-close cube. Denominator = total checker positions only
-	// (decision_type=0, same as gnuBG anTotalMoves). The DecisionType filter from
-	// the StatsFilter is deliberately ignored here so the denominator always covers
-	// both checkers and cubes (matching the gnuBG formula).
-	{
-		snowieFilter := filter
-		snowieFilter.DecisionType = -1 // count all decision types
-		snowieWhere, snowieArgs := buildBaseWhereClause(snowieFilter)
-		var snowieSumErr int64
-		var snowieCheckerCnt int
-		_ = d.db.QueryRow(
-			`SELECT COALESCE(SUM(`+statsErrExpr+`),0), `+
-				`COALESCE(SUM(CASE WHEN p.decision_type=0 THEN 1 ELSE 0 END),0) `+
-				statsBaseJoin+snowieWhere,
-			snowieArgs...,
-		).Scan(&snowieSumErr, &snowieCheckerCnt)
-		result.SnowieGlobal = snowieER(snowieSumErr, snowieCheckerCnt)
-	}
-
-	// ── 3. PR per tournament ──────────────────────────────────────────────────
-	rows, err = d.db.Query(
-		`SELECT m.tournament_id, COALESCE(t.name,''), COALESCE(t.date,''), SUM(`+statsErrExpr+`), COUNT(*) `+
-			statsBaseJoin+whereSQL+
-			` AND m.tournament_id IS NOT NULL`+
-			` GROUP BY m.tournament_id ORDER BY t.date, t.created_at`,
-		baseArgs...,
-	)
-	if err != nil {
-		return nil, fmt.Errorf("PR per tournament query: %w", err)
-	}
-	func() {
-		defer rows.Close()
-		for rows.Next() {
-			var ts TournamentStats
-			var sumErr int64
-			var cnt int
-			if err2 := rows.Scan(&ts.ID, &ts.Name, &ts.Date, &sumErr, &cnt); err2 != nil {
-				return
-			}
-			ts.NumDecisions = cnt
-			ts.PR = pr(sumErr, cnt)
-			result.PerTournament = append(result.PerTournament, ts)
-		}
-	}()
-
-	// ── 4. PR per match ───────────────────────────────────────────────────────
-	rows, err = d.db.Query(
-		`SELECT m.id, COALESCE(m.match_date,''), SUM(`+statsErrExpr+`), COUNT(*) `+
-			statsBaseJoin+whereSQL+
-			` GROUP BY m.id ORDER BY m.match_date`,
-		baseArgs...,
-	)
-	if err != nil {
-		return nil, fmt.Errorf("PR per match query: %w", err)
-	}
-	func() {
-		defer rows.Close()
-		for rows.Next() {
-			var ms MatchStats
-			var sumErr int64
-			var cnt int
-			if err2 := rows.Scan(&ms.ID, &ms.Date, &sumErr, &cnt); err2 != nil {
-				return
-			}
-			ms.NumDecisions = cnt
-			ms.PR = pr(sumErr, cnt)
-			result.PerMatch = append(result.PerMatch, ms)
-		}
-	}()
-
-	// ── 5. Cube-action breakdown ──────────────────────────────────────────────
-	{
-		cubeWhere := whereSQL + " AND p.decision_type = 1"
-		rows, err = d.db.Query(
-			`SELECT COALESCE(a.best_cube_action,''), SUM(a.cube_error), COUNT(*),`+
-				` SUM(CASE WHEN a.cube_error > ? THEN 1 ELSE 0 END) `+
-				statsBaseJoin+cubeWhere+
-				` GROUP BY a.best_cube_action`,
-			append([]any{blunderThresholdMP}, baseArgs...)...,
-		)
-		if err != nil {
-			return nil, fmt.Errorf("cube action breakdown query: %w", err)
-		}
-		func() {
-			defer rows.Close()
-			for rows.Next() {
-				var cs CubeActionStats
-				var sumErr int64
-				if err2 := rows.Scan(&cs.Action, &sumErr, &cs.NumDecisions, &cs.BlunderCount); err2 != nil {
-					return
-				}
-				cs.PR = pr(sumErr, cs.NumDecisions)
-				result.CubeActionBreakdown = append(result.CubeActionBreakdown, cs)
-			}
-		}()
-	}
-
-	// ── 6. Error histogram ────────────────────────────────────────────────────
-	histogramSQL := `SELECT
-		CASE
-			WHEN (` + statsErrExpr + `) < 5   THEN 0
-			WHEN (` + statsErrExpr + `) < 10  THEN 5
-			WHEN (` + statsErrExpr + `) < 25  THEN 10
-			WHEN (` + statsErrExpr + `) < 50  THEN 25
-			WHEN (` + statsErrExpr + `) < 100 THEN 50
-			ELSE 100
-		END as bucket,
-		COUNT(*) ` +
-		statsBaseJoin + whereSQL +
-		` GROUP BY bucket ORDER BY bucket`
-
-	rows, err = d.db.Query(histogramSQL, baseArgs...)
-	if err != nil {
-		return nil, fmt.Errorf("error histogram query: %w", err)
-	}
-	bucketMaxMap := map[int]int{0: 5, 5: 10, 10: 25, 25: 50, 50: 100, 100: -1}
-	func() {
-		defer rows.Close()
-		for rows.Next() {
-			var bucketMin, cnt int
-			if err2 := rows.Scan(&bucketMin, &cnt); err2 != nil {
-				return
-			}
-			result.ErrorHistogram = append(result.ErrorHistogram, ErrorBucket{
-				MinMP: bucketMin,
-				MaxMP: bucketMaxMap[bucketMin],
-				Count: cnt,
-			})
-		}
-	}()
-
-	// ── 7. Top blunders ───────────────────────────────────────────────────────
-	rows, err = d.db.Query(
-		`SELECT p.id, m.id, COALESCE(m.tournament_id, 0), (`+statsErrExpr+`) as emg,`+
-			` p.decision_type,`+
-			` COALESCE(m.match_date, '') as match_date,`+
-			` COALESCE(m.player1_name, '') || ' vs ' || COALESCE(m.player2_name, '') as player_names `+
-			statsBaseJoin+whereSQL+
-			` ORDER BY emg DESC LIMIT 10`,
-		baseArgs...,
-	)
-	if err != nil {
-		return nil, fmt.Errorf("top blunders query: %w", err)
-	}
-	func() {
-		defer rows.Close()
-		for rows.Next() {
-			var be BlunderEntry
-			if err2 := rows.Scan(&be.PositionID, &be.MatchID, &be.TournamentID, &be.ErrorMP,
-				&be.DecisionType, &be.MatchDate, &be.PlayerNames); err2 != nil {
-				return
-			}
-			result.TopBlunders = append(result.TopBlunders, be)
-		}
-	}()
-
-	// ── 8. Rolling PR ─────────────────────────────────────────────────────────
-	rollingNs := []int{5, 10, 50, 100, 250, 500, 1000}
-	maxN := rollingNs[len(rollingNs)-1]
-
-	recentRows, err := d.db.Query(
-		`SELECT (`+statsErrExpr+`) as err `+
-			statsBaseJoin+whereSQL+
-			` ORDER BY m.match_date DESC, mv.move_number DESC LIMIT ?`,
-		append(baseArgs, maxN)...,
-	)
-	if err != nil {
-		return nil, fmt.Errorf("rolling PR query: %w", err)
-	}
-	var recentErrors []int64
-	func() {
-		defer recentRows.Close()
-		for recentRows.Next() {
-			var e int64
-			if err2 := recentRows.Scan(&e); err2 != nil {
-				return
-			}
-			recentErrors = append(recentErrors, e)
-		}
-	}()
-
-	var cumSum int64
-	for i, e := range recentErrors {
-		cumSum += e
-		n := i + 1
-		for _, threshold := range rollingNs {
-			if n == threshold {
-				result.PRRolling[threshold] = pr(cumSum, n)
-			}
-		}
-	}
-
-	// ── MWC pass ──────────────────────────────────────────────────────────────
-	// Stream per-row data (err, position context) in most-recent-first order and
-	// aggregate MWC losses in Go. One supplementary SQL pass; O(n_decisions).
-	{
-		mwcPassSQL := `SELECT ` + statsErrExpr + ` as err,` +
-			` COALESCE(p.score_1, 0), COALESCE(p.score_2, 0), mv.player,` +
-			` (1 << COALESCE(p.cube_value, 0)), COALESCE(p.match_length, m.match_length, 0),` +
-			` COALESCE(m.tournament_id, 0), m.id,` +
-			` COALESCE(a.best_cube_action, ''), p.decision_type, p.id ` +
-			statsBaseJoin + whereSQL +
-			` ORDER BY m.match_date DESC, mv.move_number DESC`
-
-		mwcRows, mwcErr := d.db.Query(mwcPassSQL, baseArgs...)
-		if mwcErr != nil {
-			return nil, fmt.Errorf("MWC pass query: %w", mwcErr)
-		}
-
-		mwcByTournament := make(map[int64]float64)
-		mwcByMatch := make(map[int64]float64)
-		mwcByCubeAction := make(map[string]float64)
-		blunderMWC := make(map[int64]float64)
-
-		var mwcGlobal, mwcChecker, mwcCube float64
-		var mwcAvailable bool
-		var rowIdx int
-		var mwcRollingCum float64
-		mwcRollingThresholds := []int{5, 10, 50, 100, 250, 500, 1000}
-		mwcRollingMap := make(map[int]float64)
-
-		func() {
-			defer mwcRows.Close()
-			for mwcRows.Next() {
-				var errMP int64
-				var awayScore0, awayScore1, rawPlayer, cubeValue, matchLength int
-				var tournamentID, matchID int64
-				var cubeAction string
-				var dt int
-				var posID int64
-				if err2 := mwcRows.Scan(&errMP, &awayScore0, &awayScore1, &rawPlayer, &cubeValue, &matchLength,
-					&tournamentID, &matchID, &cubeAction, &dt, &posID); err2 != nil {
-					return
-				}
-
-				rowIdx++
-
-				// Convert XG player encoding (1 or -1) to gnuBG fMove (0 or 1).
-				// XG encodes player 0 (bottom) as 1 and player 1 (top) as -1.
-				fMove := 0
-				if rawPlayer == -1 {
-					fMove = 1
-				}
-				// p.score_1/score_2 are away scores (points still needed); gnuBGGetME
-				// expects current scores (games already won).
-				currentScore0 := matchLength - awayScore0
-				currentScore1 := matchLength - awayScore1
-
-				mwcLoss := ConvertEMGLossToMWCLoss(int(errMP), currentScore0, currentScore1, fMove, cubeValue, matchLength)
-
-				if !math.IsNaN(mwcLoss) {
-					mwcAvailable = true
-					mwcGlobal += mwcLoss
-					if dt == 0 {
-						mwcChecker += mwcLoss
-					} else {
-						mwcCube += mwcLoss
-					}
-					if tournamentID != 0 {
-						mwcByTournament[tournamentID] += mwcLoss
-					}
-					mwcByMatch[matchID] += mwcLoss
-					if dt == 1 {
-						mwcByCubeAction[cubeAction] += mwcLoss
-					}
-					blunderMWC[posID] = mwcLoss
-					mwcRollingCum += mwcLoss
-				}
-
-				// Rolling thresholds count all decisions (including money-game), mirroring PRRolling.
-				for _, threshold := range mwcRollingThresholds {
-					if rowIdx == threshold {
-						mwcRollingMap[threshold] = mwcRollingCum
-					}
-				}
-			}
-		}()
-
-		result.MWCGlobal = mwcGlobal
-		result.MWCChecker = mwcChecker
-		result.MWCCube = mwcCube
-		result.MWCAvailable = mwcAvailable
-		result.MWCRolling = mwcRollingMap
-
-		for i, ts := range result.PerTournament {
-			result.PerTournament[i].MWC = mwcByTournament[ts.ID]
-		}
-		for i, ms := range result.PerMatch {
-			result.PerMatch[i].MWC = mwcByMatch[ms.ID]
-		}
-		for i, cs := range result.CubeActionBreakdown {
-			result.CubeActionBreakdown[i].MWC = mwcByCubeAction[cs.Action]
-		}
-		for i, be := range result.TopBlunders {
-			if loss, ok := blunderMWC[be.PositionID]; ok {
-				result.TopBlunders[i].MWCLoss = loss
-			}
-		}
-	}
-
-	return result, nil
-}
-
-// ── Drill-down ────────────────────────────────────────────────────────────────
 
 // SelectionSpec describes what the user selected in the stats panel (a click on
 // a bar, point, or row). The frontend passes this to GetPositionIDsByStatsSelection
@@ -645,81 +139,6 @@ type SelectionSpec struct {
 	OnlyWithError bool // for "cube_action", "checker", "cube" → error > 0
 }
 
-// buildSelectionWhereClause produces the extra WHERE fragment and optional
-// ORDER BY / LIMIT fragment for a given SelectionSpec.
-//
-// whereAdd starts with " AND " or is empty (never contains "WHERE").
-// orderLimit is the ORDER BY … LIMIT … suffix or empty.
-// args contains the combined parameters for whereAdd followed by orderLimit.
-func buildSelectionWhereClause(sel SelectionSpec) (whereAdd string, orderLimit string, args []any) {
-	switch sel.Kind {
-	case "checker":
-		whereAdd = " AND p.decision_type = 0"
-		if sel.OnlyWithError {
-			whereAdd += " AND (" + statsErrExpr + ") > 0"
-		}
-	case "cube":
-		whereAdd = " AND p.decision_type = 1"
-		if sel.OnlyWithError {
-			whereAdd += " AND (" + statsErrExpr + ") > 0"
-		}
-	case "cube_action":
-		whereAdd = " AND p.decision_type = 1 AND a.best_cube_action = ?"
-		args = append(args, sel.CubeAction)
-		if sel.OnlyWithError {
-			whereAdd += " AND (" + statsErrExpr + ") > 0"
-		}
-	case "error_bucket":
-		whereAdd = " AND (" + statsErrExpr + ") >= ?"
-		args = append(args, sel.BucketMinMP)
-		if sel.BucketMaxMP != -1 {
-			whereAdd += " AND (" + statsErrExpr + ") < ?"
-			args = append(args, sel.BucketMaxMP)
-		}
-	case "tournament":
-		whereAdd = " AND m.tournament_id = ?"
-		args = append(args, sel.TournamentID)
-	case "match":
-		whereAdd = " AND m.id = ?"
-		args = append(args, sel.MatchID)
-	case "last_n":
-		orderLimit = "ORDER BY m.match_date DESC, mv.move_number DESC LIMIT ?"
-		args = append(args, sel.LastN)
-	case "position":
-		whereAdd = " AND p.id = ?"
-		args = append(args, sel.PositionID)
-	case "top_blunders":
-		limit := 10
-		if sel.LastN > 0 {
-			limit = sel.LastN
-		}
-		orderLimit = "ORDER BY (" + statsErrExpr + ") DESC LIMIT ?"
-		args = append(args, limit)
-		// "all" → no extra clauses
-	}
-	return whereAdd, orderLimit, args
-}
-
-// scanPositionIDs scans a rows result set returning a single int64 column into
-// a slice. It closes rows and propagates any iteration error.
-func scanPositionIDs(rows interface {
-	Next() bool
-	Scan(...any) error
-	Close() error
-	Err() error
-}) ([]int64, error) {
-	defer rows.Close()
-	var ids []int64
-	for rows.Next() {
-		var id int64
-		if err := rows.Scan(&id); err != nil {
-			return nil, fmt.Errorf("scan position id: %w", err)
-		}
-		ids = append(ids, id)
-	}
-	return ids, rows.Err()
-}
-
 // GetPositionIDsByStatsSelection resolves a user selection made in the Stats
 // panel into a deduplicated list of position IDs. The StatsFilter is always
 // applied so that the IDs correspond exactly to what is displayed in the panel
@@ -729,30 +148,6 @@ func (d *Database) GetPositionIDsByStatsSelection(filter StatsFilter, sel Select
 	d.mu.RLock()
 	defer d.mu.RUnlock()
 	return d.store.Stats().PositionIDsBySelection(context.Background(), "", toStorageStatsFilter(filter), toStorageSelectionSpec(sel))
-}
-
-func legacyGetPositionIDsByStatsSelection(d *Database, filter StatsFilter, sel SelectionSpec) ([]int64, error) {
-	d.mu.RLock()
-	defer d.mu.RUnlock()
-
-	if d.db == nil {
-		return nil, fmt.Errorf("no database is currently open")
-	}
-
-	whereSQL, baseArgs := buildStatsWhereClause(filter)
-	whereAdd, orderLimit, selArgs := buildSelectionWhereClause(sel)
-
-	query := "SELECT DISTINCT p.id " + statsBaseJoin + whereSQL + whereAdd
-	if orderLimit != "" {
-		query += " " + orderLimit
-	}
-
-	allArgs := append(append([]any{}, baseArgs...), selArgs...)
-	rows, err := d.db.Query(query, allArgs...)
-	if err != nil {
-		return nil, fmt.Errorf("GetPositionIDsByStatsSelection (%s): %w", sel.Kind, err)
-	}
-	return scanPositionIDs(rows)
 }
 
 // GetPositionIDsByTournament returns all position IDs belonging to the given
@@ -765,23 +160,6 @@ func (d *Database) GetPositionIDsByTournament(tournamentID int64) ([]int64, erro
 	return d.store.Stats().PositionIDsByTournament(context.Background(), "", tournamentID)
 }
 
-func legacyGetPositionIDsByTournament(d *Database, tournamentID int64) ([]int64, error) {
-	d.mu.RLock()
-	defer d.mu.RUnlock()
-
-	if d.db == nil {
-		return nil, fmt.Errorf("no database is currently open")
-	}
-
-	query := "SELECT DISTINCT p.id " + statsBaseJoin +
-		" WHERE m.tournament_id = ? AND a.position_id IS NOT NULL AND (" + statsErrExpr + ") IS NOT NULL"
-	rows, err := d.db.Query(query, tournamentID)
-	if err != nil {
-		return nil, fmt.Errorf("GetPositionIDsByTournament: %w", err)
-	}
-	return scanPositionIDs(rows)
-}
-
 // GetPositionIDsByMatch returns all position IDs belonging to the given match,
 // regardless of any stats filter. Used when the user explicitly reopens a match.
 // GetPositionIDsByMatch delegates to the storage StatsStore.
@@ -789,23 +167,6 @@ func (d *Database) GetPositionIDsByMatch(matchID int64) ([]int64, error) {
 	d.mu.RLock()
 	defer d.mu.RUnlock()
 	return d.store.Stats().PositionIDsByMatch(context.Background(), "", matchID)
-}
-
-func legacyGetPositionIDsByMatch(d *Database, matchID int64) ([]int64, error) {
-	d.mu.RLock()
-	defer d.mu.RUnlock()
-
-	if d.db == nil {
-		return nil, fmt.Errorf("no database is currently open")
-	}
-
-	query := "SELECT DISTINCT p.id " + statsBaseJoin +
-		" WHERE m.id = ? AND a.position_id IS NOT NULL AND (" + statsErrExpr + ") IS NOT NULL"
-	rows, err := d.db.Query(query, matchID)
-	if err != nil {
-		return nil, fmt.Errorf("GetPositionIDsByMatch: %w", err)
-	}
-	return scanPositionIDs(rows)
 }
 
 // PlayerFrequency pairs a player name with the number of matches in which they appear.
@@ -826,40 +187,6 @@ func (d *Database) GetAllPlayerNames() ([]PlayerFrequency, error) {
 		return nil, err
 	}
 	return fromStoragePlayerFreq(p), nil
-}
-
-func legacyGetAllPlayerNames(d *Database) ([]PlayerFrequency, error) {
-	d.mu.RLock()
-	defer d.mu.RUnlock()
-
-	if d.db == nil {
-		return nil, fmt.Errorf("no database is currently open")
-	}
-
-	rows, err := d.db.Query(`
-		SELECT name, COUNT(*) AS cnt
-		FROM (
-			SELECT player1_name AS name FROM match WHERE player1_name != ''
-			UNION ALL
-			SELECT player2_name AS name FROM match WHERE player2_name != ''
-		)
-		GROUP BY name
-		ORDER BY cnt DESC, name ASC
-	`)
-	if err != nil {
-		return nil, fmt.Errorf("GetAllPlayerNames: %w", err)
-	}
-	defer rows.Close()
-
-	var result []PlayerFrequency
-	for rows.Next() {
-		var pf PlayerFrequency
-		if err := rows.Scan(&pf.Name, &pf.Count); err != nil {
-			return nil, fmt.Errorf("GetAllPlayerNames scan: %w", err)
-		}
-		result = append(result, pf)
-	}
-	return result, rows.Err()
 }
 
 // applyMatchBadges fills the per-player PR/MWC badge fields of every match in
@@ -963,217 +290,4 @@ func (d *Database) GetMatchDetailStats(matchID int64) (*MatchDetailStats, error)
 		return nil, err
 	}
 	return fromStorageMatchDetail(m), nil
-}
-
-func legacyGetMatchDetailStats(d *Database, matchID int64) (*MatchDetailStats, error) {
-	d.mu.RLock()
-	defer d.mu.RUnlock()
-
-	if d.db == nil {
-		return nil, fmt.Errorf("no database is currently open")
-	}
-
-	query := `SELECT mv.player, p.decision_type, COALESCE(mv.cube_action,''),
-		(` + statsErrExpr + `) as err_mp,
-		COALESCE(p.score_1, 0), COALESCE(p.score_2, 0),
-		(1 << COALESCE(p.cube_value, 0)),
-		COALESCE(p.match_length, m.match_length, 0) ` +
-		statsBaseJoin +
-		` WHERE m.id = ? AND a.position_id IS NOT NULL AND (` + statsErrExpr + `) IS NOT NULL AND ` + statsCountedExpr
-
-	rows, err := d.db.Query(query, matchID)
-	if err != nil {
-		return nil, fmt.Errorf("GetMatchDetailStats query: %w", err)
-	}
-	defer rows.Close()
-
-	type playerAcc struct {
-		totalSumErr   int64
-		totalCnt      int
-		totalErrors   int
-		totalBlunders int
-		totalMWC      float64
-
-		checkerSumErr   int64
-		checkerCnt      int
-		checkerErrors   int
-		checkerBlunders int
-		checkerMWC      float64
-
-		doubleSumErr   int64
-		doubleCnt      int
-		doubleErrors   int
-		doubleBlunders int
-		doubleMWC      float64
-
-		takeSumErr   int64
-		takeCnt      int
-		takeErrors   int
-		takeBlunders int
-		takeMWC      float64
-	}
-
-	var p1, p2 playerAcc
-
-	for rows.Next() {
-		var rawPlayer, decisionType int
-		var cubeAction string
-		var errMP int64
-		var awayScore0, awayScore1, cubeValue, matchLength int
-		if err := rows.Scan(&rawPlayer, &decisionType, &cubeAction, &errMP,
-			&awayScore0, &awayScore1, &cubeValue, &matchLength); err != nil {
-			continue
-		}
-
-		fMove := 0
-		if rawPlayer == -1 {
-			fMove = 1
-		}
-		// p.score_1/score_2 are away scores; ConvertEMGLossToMWCLoss expects current scores.
-		currentScore0 := matchLength - awayScore0
-		currentScore1 := matchLength - awayScore1
-		mwcLoss := ConvertEMGLossToMWCLoss(int(errMP), currentScore0, currentScore1, fMove, cubeValue, matchLength)
-		if math.IsNaN(mwcLoss) {
-			mwcLoss = 0
-		}
-
-		isError := errMP > 0
-		isBlunder := errMP >= blunderThresholdMP
-		isTake := cubeAction == "Take" || cubeAction == "Pass"
-
-		acc := &p1
-		if rawPlayer == -1 {
-			acc = &p2
-		}
-
-		acc.totalSumErr += errMP
-		acc.totalCnt++
-		acc.totalMWC += mwcLoss
-		if isError {
-			acc.totalErrors++
-		}
-		if isBlunder {
-			acc.totalBlunders++
-		}
-
-		if decisionType == 0 {
-			acc.checkerSumErr += errMP
-			acc.checkerCnt++
-			acc.checkerMWC += mwcLoss
-			if isError {
-				acc.checkerErrors++
-			}
-			if isBlunder {
-				acc.checkerBlunders++
-			}
-		} else {
-			if isTake {
-				acc.takeSumErr += errMP
-				acc.takeCnt++
-				acc.takeMWC += mwcLoss
-				if isError {
-					acc.takeErrors++
-				}
-				if isBlunder {
-					acc.takeBlunders++
-				}
-			} else {
-				acc.doubleSumErr += errMP
-				acc.doubleCnt++
-				acc.doubleMWC += mwcLoss
-				if isError {
-					acc.doubleErrors++
-				}
-				if isBlunder {
-					acc.doubleBlunders++
-				}
-			}
-		}
-	}
-	if err := rows.Err(); err != nil {
-		return nil, fmt.Errorf("GetMatchDetailStats scan: %w", err)
-	}
-
-	// ── Snowie ER pass ────────────────────────────────────────────────────────
-	// Second query without statsCountedExpr: sum all equity errors per player,
-	// count checker positions (decision_type=0, forced included) for both players.
-	// Denominator = anTotalMoves[P1] + anTotalMoves[P2] (gnuBG convention).
-	var snowieP1SumErr, snowieP2SumErr int64
-	var snowieP1Checker, snowieP2Checker int
-	{
-		snowieRows, snowieErr := d.db.Query(
-			`SELECT mv.player, p.decision_type, (`+statsErrExpr+`) as err_mp `+
-				statsBaseJoin+
-				` WHERE m.id = ? AND a.position_id IS NOT NULL AND (`+statsErrExpr+`) IS NOT NULL`,
-			matchID)
-		if snowieErr != nil {
-			return nil, fmt.Errorf("GetMatchDetailStats snowie query: %w", snowieErr)
-		}
-		func() {
-			defer snowieRows.Close()
-			for snowieRows.Next() {
-				var rawPlayer, decisionType int
-				var errMP int64
-				if err2 := snowieRows.Scan(&rawPlayer, &decisionType, &errMP); err2 != nil {
-					return
-				}
-				if rawPlayer == 1 {
-					snowieP1SumErr += errMP
-					if decisionType == 0 {
-						snowieP1Checker++
-					}
-				} else {
-					snowieP2SumErr += errMP
-					if decisionType == 0 {
-						snowieP2Checker++
-					}
-				}
-			}
-		}()
-	}
-	snowieDenom := snowieP1Checker + snowieP2Checker
-
-	buildStats := func(a *playerAcc) MatchPlayerDetailStats {
-		cubeCnt := a.doubleCnt + a.takeCnt
-		cubeSumErr := a.doubleSumErr + a.takeSumErr
-		return MatchPlayerDetailStats{
-			TotalDecisions:   a.totalCnt,
-			TotalErrors:      a.totalErrors,
-			TotalBlunders:    a.totalBlunders,
-			TotalEquityError: float64(a.totalSumErr) / 1000,
-			PR:               pr(a.totalSumErr, a.totalCnt),
-			MWCLoss:          a.totalMWC,
-
-			CheckerDecisions:   a.checkerCnt,
-			CheckerErrors:      a.checkerErrors,
-			CheckerBlunders:    a.checkerBlunders,
-			CheckerEquityError: float64(a.checkerSumErr) / 1000,
-			PRChecker:          pr(a.checkerSumErr, a.checkerCnt),
-			CheckerMWCLoss:     a.checkerMWC,
-
-			DoubleDecisions:   a.doubleCnt,
-			DoubleErrors:      a.doubleErrors,
-			DoubleBlunders:    a.doubleBlunders,
-			DoubleEquityError: float64(a.doubleSumErr) / 1000,
-			DoubleMWCLoss:     a.doubleMWC,
-
-			TakeDecisions:   a.takeCnt,
-			TakeErrors:      a.takeErrors,
-			TakeBlunders:    a.takeBlunders,
-			TakeEquityError: float64(a.takeSumErr) / 1000,
-			TakeMWCLoss:     a.takeMWC,
-
-			PRCube:      pr(cubeSumErr, cubeCnt),
-			CubeMWCLoss: a.doubleMWC + a.takeMWC,
-		}
-	}
-
-	stats := &MatchDetailStats{
-		MatchID: matchID,
-		Player1: buildStats(&p1),
-		Player2: buildStats(&p2),
-	}
-	stats.Player1.SnowieER = snowieER(snowieP1SumErr, snowieDenom)
-	stats.Player2.SnowieER = snowieER(snowieP2SumErr, snowieDenom)
-	return stats, nil
 }
